[backend,fclient] Save a map of users
Retrieve all users from the server and save off to the map. UserInfo
gives information if the user even exists at all while UserMap under
ServerInfo specifies if there is user data for a user in a specific
server

This is so we can create users and userdata on demand (depending on
usage) instead of creating users for everyone in every server.